+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Add `signatureVerify (message: Uint8Array | string, signature: Uint8Array | string, addressOrPublicKey: Uint8Array | string): VerifyResult` to verify any signature, regardless of type. Returns validity & detected crypto type.
 - Add `cryptoIsReady (): boolean` function to check status of initialization
 - Add `addressCheckChecksum (decoded: Uint8Array)` as an helper to extract and verify the ss58 checksum
-- Swap to yarn 2, allow use in yan 2 projects
+- Swap to yarn 2, allow use in yarn 2 projects
 
 # 2.5.1 Feb 22, 2020